feat: refactor application page, adds mobile responsive to application screen
