+++ components/Layout/Layout.js
@@ -10,7 +10,7 @@ export default function Layout({ title = "Hyperscale", children }) {
         <meta name="viewport" content="initial-scale=1.0, width=device-width" />
       </Head>
       <div className="flex justify-center">
-        <div className="w-screen max-w-7xl">
+        <div className="w-screen max-w-7xl px-4 xl:px-0">
           <Header />
           {children}
         </div>
